refactor: use one class for template method implementation
