+++ packages/backend/.eslintrc.json
@@ -73,10 +73,12 @@
                         }
                     }
                 ],
+                "no-unused-vars":"off",
                 "@typescript-eslint/no-unused-vars": [
                     "error",
                     {
-                        "varsIgnorePattern": "^_.*$"
+                        "varsIgnorePattern": "^_",
+                        "argsIgnorePattern": "^_"
                     }
                 ],
                 "@typescript-eslint/space-before-function-paren": [
