+++ packages/backend/src/app/app-connection/app-connection-service/index.ts
@@ -1,4 +1,3 @@
 import { AppConnectionService } from './app-connection-service'
-import { DefaultAppConnectionService } from './default-app-connection-service'
 
-export const appConnectionService: AppConnectionService = new DefaultAppConnectionService()
+export const appConnectionService = new AppConnectionService()
